server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0
websockets==11.0.2
uuid==1.30
